requirements.txt
Django==1.11
djangorestframework==3.10.2
pytz==2019.2
